chore: split signal handler in log init
